+++ docs/examples/00-describe/05-csv-3-files.php
@@ -4,7 +4,15 @@ require_once __DIR__ . "/.init.php";
 
 $chainProcessor = getChainProcessor(__FILE__);
 
+$output = new \Symfony\Component\Console\Output\ConsoleOutput();
+$symfonyOutput = new \Oliverde8\Component\PhpEtl\Output\SymfonyConsoleOutput($output);
+
 $chainProcessor->process(
-    new ArrayIterator([__DIR__ . "/customers.csv", __DIR__ . "/customers2.csv"]),
-    []
+    new ArrayIterator(["/customers.csv", "/customers2.csv"]),
+    [],
+    function (array $operationStates, int $itemsProcessed, int $itemsReturned, bool $hasEnded) use ($symfonyOutput) {
+        $symfonyOutput->output($operationStates, $hasEnded);
+        // Just so that the output is nicer.
+        usleep(100000);
+    }
 );