+++ src/Oliverde8/Component/PhpEtl/ChainObserver/ChainObserver.php
@@ -11,6 +11,8 @@ class ChainObserver extends OperationState implements ChainObserverInterface
     /** @var OperationState[] */
     protected array $operationStates = [];
 
+    protected bool $ended = false;
+
     protected $callable;
 
     /**
@@ -25,8 +27,8 @@ class ChainObserver extends OperationState implements ChainObserverInterface
 
     public function init(array $chainLinks, array $chainNames): void
     {
-        foreach ($chainLinks as $id => $opeartion) {
-            $this->operationStates[$id] = new OperationState($chainNames[$id]);
+        foreach ($chainLinks as $id => $operation) {
+            $this->operationStates[$id] = new OperationState($chainNames[$id], $operation);
         }
 
         $this->callback();
@@ -34,23 +36,40 @@ class ChainObserver extends OperationState implements ChainObserverInterface
 
     public function onBeforeProcess($operationId, ChainOperationInterface $operation, ItemInterface $item): void
     {
-        $this->processItem($item);
-        $this->operationStates[$operationId]->processItem($item);
+        $this->ended = false;
+        $this->processItem($operation, $item);
+        $this->operationStates[$operationId]->processItem($operation, $item);
 
         $this->callback();
     }
 
     public function onAfterProcess($operationId, ChainOperationInterface $operation, ItemInterface $item): void
     {
-        $this->returnItem($item);
-        $this->operationStates[$operationId]->returnItem($item);
+        $this->returnItem($operation, $item);
+        $this->operationStates[$operationId]->returnItem($operation, $item);
 
         $this->callback();
+    }
+
+    public function onFinish()
+    {
+        $this->ended = true;
+        $this->callback();
+    }
+
+    public function isFinished(): bool
+    {
+        return $this->ended;
     }
 
     private function callback()
     {
         $callback = $this->callable;
-        $callback($this->operationStates, $this->getItemsProcessed(), $this->getItemsReturned());
+        $callback($this->operationStates, $this->getItemsProcessed(), $this->getItemsReturned(), $this->ended);
+    }
+
+    public function getOperationStates(): array
+    {
+        return $this->operationStates;
     }
 }