+++ src/Oliverde8/Component/PhpEtl/ChainObserver/ChainObserverInterface.php
@@ -11,5 +11,8 @@ interface ChainObserverInterface
 
     public function onBeforeProcess($operationId, ChainOperationInterface $operation, ItemInterface $item): void;
 
-    public function onAfterProcess($chainId, ChainOperationInterface $operation, ItemInterface $item): void;
+    public function onAfterProcess($operationId, ChainOperationInterface $operation, ItemInterface $item): void;
+
+    public function onFinish();
 }
+