+++ src/Oliverde8/Component/PhpEtl/Item/StopItem.php
@@ -13,5 +13,5 @@ namespace Oliverde8\Component\PhpEtl\Item;
  */
 class StopItem implements ItemInterface
 {
-
+    public function __construct(public readonly bool $isFinal = false){}
 }