+++ src/Oliverde8/Component/PhpEtl/Model/File/AbstractCsvFile.php
@@ -29,12 +29,12 @@ class AbstractCsvFile
     /**
      * Reader constructor.
      *
-     * @param string $filePath  Path to the csv file.
+     * @param string|resource $filePath  Path to the csv file.
      * @param string $delimiter Delimiter to use for the csv file.
      * @param string $enclosure Enclosure to use for the csv file.
      * @param string $escape    Escape to use for the csv file.
      */
-    public function __construct($filePath, $delimiter = ';', $enclosure = '"', $escape = '\\')
+    public function __construct($filePath, string $delimiter = ';', string $enclosure = '"', string $escape = '\\')
     {
         $this->filePath = $filePath;
         $this->delimiter = $delimiter;