+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # 1.2.0
 - :star2: Feature #14 - Added possibility to create subchains.
+- :star2: Feature #34 - Allow chain's to be observed to see progress.
+- :star2: Feature #35 - Added a symfony console output to display nice output based on the chain's progress.
+- :star2: Feature #42 - Added output for generating mermaid graph for a etl chain definition & for running chain.
+- :wrench: Fix - Boolean false values being converted to string by the rule engine.
+- :wrench: Fix - Http client operation not handling empty response properly when response is json.
+
+# 1.1.4
+- :star2: Split item operation can now return more complex data sets using the `keyName` option
+- :star2: Split item operation can now keep/duplicate data before splitting it using the `duplicateKeys` option.
+- :star2: Split item operation can now keep the original keys while splitting it using the `keepKeys` option.
+
+# 1.1.3
+- :wrench: Fix - Added support for symfony 7
 
 # 1.1.2
 - :wrench: Fix - Usage of the AbstractFactory for operation without overriding configureValidator causing errors.